Python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = ""
authors = [
    {name = "r3d5un",email = "r3d5un@pm.me"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "structlog (>=25.2.0,<26.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "testcontainers[postgres] (>=4.10.0,<5.0.0)",
    "psycopg-binary (>=3.2.6,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.53b1,<0.54)",
    "opentelemetry-api (>=1.32.1,<2.0.0)",
    "opentelemetry-sdk (>=1.32.1,<2.0.0)",
    "opentelemetry-distro (>=0.53b1,<0.54)",
]

[tool.poetry]
packages = [{include = "python", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
